requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
tiktoken>=0.5.0
pydantic>=2.5.0
jinja2>=3.1.0
aiofiles>=23.2.0
# v2.0 New
langgraph>=0.2.0
langsmith>=0.1.0
